Create debonding delegation list component
